+++ aider/models.py
@@ -76,6 +76,7 @@ MODEL_ALIASES = {
     "3": "gpt-3.5-turbo",
     # Other models
     "deepseek": "deepseek/deepseek-chat",
+    "r1": "deepseek/deepseek-reasoner",
     "flash": "gemini/gemini-2.0-flash-exp",
 }
 
@@ -621,6 +622,21 @@ MODEL_SETTINGS = [
         "diff",
         use_repo_map=True,
         send_undo_reply=False,
+    ),
+    ModelSettings(
+        "openrouter/deepseek/deepseek-r1",
+        "diff",
+        weak_model_name="openrouter/deepseek/deepseek-chat",
+        editor_model_name="openrouter/deepseek/deepseek-chat",
+        editor_edit_format="editor-diff",
+        use_repo_map=True,
+        examples_as_sys_msg=True,
+        use_temperature=False,
+        reminder="user",
+        caches_by_default=True,
+        extra_params={
+            "max_tokens": 8192,
+        },
     ),
     ModelSettings(
         "deepseek/deepseek-reasoner",